change from servo.h to direct analogWrite

+++ Robot.h
@@ -28,7 +28,7 @@ class Robot {
 		void init();
 		Ecran _ecran;
 		BNOSensor _sensor;
-		void directMove(byte servo, int angle);
+		void directMove(byte servo, byte angle);
 	private:
 		MyServo servos[NOMBRE_SERVO];
 };
